chatting function 100%

+++ backend/src/main/java/com/matching/system/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.matching.system.service;
 
 import com.matching.system.domain.Category;
+import com.matching.system.domain.MatchingPost;
 import com.matching.system.dto.CategoryDTO;
 import com.matching.system.process.ImageProcess;
 import com.matching.system.response.ResponseData;
@@ -69,8 +70,14 @@ public class CategoryService {
     public ResponseMessage delete(Long categoryId)
     {
         // 연관 post에는 null
-        matchingPostRepository.findByCategoryId(categoryId).stream()
-                .forEach(matchingPost -> matchingPost.deleteCategory());
+//        matchingPostRepository.findByCategoryId(categoryId).stream()
+//                .forEach(matchingPost -> matchingPost.deleteCategory());
+        Category findCategory = categoryRepository.findById(categoryId).get();
+
+        Optional<MatchingPost> findMatchingPost = matchingPostRepository.findByExistCategory(findCategory);
+        if (! findMatchingPost.isEmpty())
+            return new ResponseMessage(HttpStatus.NOT_ACCEPTABLE, "이미 이 카테고리는 사용중입니다.");
+
 
         categoryRepository.deleteById(categoryId);
